Creating CityCell File
Creating CityCell file to separate from HomeView
